remove Navigator interfaces to their own file
